refactor plugin to be more easily usable as a lib
